+++ lib/src/merge_requests.dart
@@ -33,7 +33,7 @@ class MergeRequestsApi {
     final uri = _project.buildUri(['merge_requests'],
         queryParameters: queryParameters, page: page, perPage: perPage);
 
-    final List<Map> jsonList = await _gitLab.request(uri);
+    final jsonList = _responseToList(await _gitLab.request(uri));
 
     return jsonList.map((json) => new MergeRequest.fromJson(json)).toList();
   }