+++ test/merge_requests_test.dart
@@ -6,6 +6,7 @@ import 'package:gitlab/gitlab.dart';
 import 'package:mockito/mockito.dart';
 import 'package:test/test.dart';
 
+import 'assets/json_data.dart' as data;
 import 'src/mocks.dart';
 
 void main() {
@@ -18,15 +19,17 @@ void main() {
     final projectId = 1337;
     final headers = {'PRIVATE-TOKEN': 'secret-token'};
 
-    final mergeRequestsJson =
-        new File('test/assets/merge_requests.get.json').readAsStringSync();
-    final Map mergeRequestsMap = jsonDecode(mergeRequestsJson);
+    final mergeRequestsMap = data.decodeMap(data.mergeRequest);
     final int mergeRequestId = mergeRequestsMap['id'];
     final int mergeRequestIid = mergeRequestsMap['iid'];
 
     setUp(() {
       mockResponse = new MockResponse();
       mockHttpClient = new MockGitLabHttpClient();
+
+      when(mockHttpClient.request(any, any, any))
+          .thenAnswer((_) => new Future.value(mockResponse));
+
       gitLab = getTestable(mockHttpClient);
       project = gitLab.project(projectId);
     });
@@ -57,13 +60,20 @@ void main() {
     test('.get()', () async {
       final uri = Uri.parse('https://gitlab.com/api/v4'
           '/projects/$projectId/merge_requests/$mergeRequestIid');
-      when(mockHttpClient.request(uri, headers, HttpMethod.get))
-          .thenAnswer((_) => new Future.value(mockResponse));
       when(mockResponse.statusCode).thenReturn(200);
-      when(mockResponse.body).thenReturn(mergeRequestsJson);
+      when(mockResponse.body).thenReturn(data.mergeRequest);
       final mergeRequest = await project.mergeRequests.get(mergeRequestId);
       verify(mockHttpClient.request(uri, headers, HttpMethod.get)).called(1);
       expect(mergeRequest.id, mergeRequestId);
+    });
+    test('.list()', () async {
+      final uri = Uri.parse('https://gitlab.com/api/v4'
+          '/projects/$projectId/merge_requests?');
+      when(mockResponse.statusCode).thenReturn(200);
+      when(mockResponse.body).thenReturn(data.mergeRequests);
+      final mergeRequests = await project.mergeRequests.list();
+      verify(mockHttpClient.request(uri, headers, HttpMethod.get)).called(1);
+      expect(mergeRequests, hasLength(1));
     });
   });
 }